Add ExceptionHandler and modify few files

+++ models.py
@@ -1,11 +1,12 @@
 from sqlmodel import Field, SQLModel
 from datetime import datetime 
 
+
 class Requests(SQLModel, table=True):
     id: int = Field(default=None, nullable=False, primary_key=True)
     time: datetime = Field(default=datetime.now())
     stock: str 
-    price: int 
+    price: float
     av_7: float 
     av_14: float
     av_21: float
